fix(query-intelligence): PR review fixes — migration head, unsound rewrite, guards
Review fixes for the query-intelligence feature branch:

- alembic: re-parent qryintel01 onto mrgperf01 — after merging main it
  shared down_revision ff8803de3eb2 with the mrgperf01 merge migration,
  leaving two heads and breaking 'alembic upgrade head'. Add a unit test
  that walks alembic/versions and asserts a single head.
- sql_optimizer: the ORDER BY-stripping rewrite was not semantics-
  preserving when an order-sensitive aggregate (array_agg, string_agg,
  group_concat, listagg, ...) consumes the subquery's ordering, or when
  the subquery uses DISTINCT (Postgres DISTINCT ON keeps rows by ORDER
  BY). Both cases now leave the SQL untouched.
- profiler: honor the profile_max_table_rows guard — drop value
  dictionaries when the table's row count exceeds it, since a bounded
  sample of a huge table can't enumerate every category and would cause
  false-positive 'matches no observed value' lints.
- table_profile: row_count_estimate Integer -> BigInteger (warehouse
  tables routinely exceed int32).
- cost_recorder: compute summarize_timings once per table instead of
  twice in attribute_costs_to_tables.
- routes: profiling endpoint now requires 'manage' (it writes profile
  rows and issues sampling queries) and validates table_limit >= 1.
- code_execution: the lake path now records optimizer notes/warnings on
  its timing entries like the eager path.
- tests: cover the new rewrite guards, the huge-table dictionary guard,
  and cache-hit-only cost merges; update the stale ms_fabric OAuth scope
  assertion to the intentional database.windows.net scope.

+++ backend/app/models/table_profile.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, String, ForeignKey, JSON, Integer, DateTime, Index
+from sqlalchemy import Column, String, ForeignKey, JSON, Integer, BigInteger, DateTime, Index
 from sqlalchemy.orm import relationship
 
 from app.models.base import BaseSchema
@@ -38,7 +38,8 @@ class TableProfile(BaseSchema):
     table_fqn = Column(String, nullable=False)
 
     # Best-effort row count from the sampling pass (COUNT(*) or an approx).
-    row_count_estimate = Column(Integer, nullable=True)
+    # BigInteger: real warehouse tables routinely exceed the int32 range.
+    row_count_estimate = Column(BigInteger, nullable=True)
     # Number of rows actually sampled to derive the column stats below.
     sample_rows = Column(Integer, nullable=False, default=0)
